Fix a pytest error and move dummy test location
Fix pytest unknown module error and move the dummy test into
student\tests.py
